+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Guide: https://keepachangelog.com/en/1.0.0/
 ### Breaking changes
 - [CORE] `SearchEngine.setAccessToken(_: String)` has been removed.
 
+
+### Fixed
+- [UI]: fixed UIViewController presentation on SDK side. [#7 ](https://github.com/mapbox/mapbox-search-ios/issues/7)
+
 ## [1.0.0-beta.33] - 2022-07-20
 
 ### Breaking changes